gexray2_db_to_sse/pyproject.toml
[project]
name = "gex2_to_db"
version = "0.0.3"
description = "Docker container for legacy Gex api."
requires-python = ">=3.12"
dependencies = [
    "asyncpg==0.30.0",
    "python-dotenv==1.0.1",
    "redis==5.2.1",
    "hiredis==3.1.0",
]
